Initial draft conversion of netlogo-headless

+++ netlogo-headless/src/main/prim/_randomconst.scala
@@ -10,5 +10,5 @@ class _randomconst(n: Long) extends Reporter {
   override def report(context: Context): java.lang.Double =
     Double.box(report_1(context))
   def report_1(context: Context): Double =
-    context.job.random.nextLong(n)
+    context.job.random.nextLong(n).toDouble
 }
